Bugfix - beregning av feilutbetaling
